+++ module/utility/upload_hdc.js
@@ -353,20 +353,42 @@ export async function applyCharacterSheet(xmlDoc) {
     // combat maneuvers
     async function loadCombatManeuvers(dict, actor) {
         for (const entry of Object.entries(dict)) {
-            const v = entry[1]
+            const name = entry[0];
+            const v = entry[1];
+            const phase = v[0];
+            const ocv = v[1];
+            const dcv = v[2];
+            const effects = v[3];
+            const attack = v[4];
+            const XMLID = name.toUpperCase().replace(" ", ""); // A fake XMLID
             const itemData = {
-                name: entry[0],
+                name,
                 type: 'maneuver',
-                data: {
-                    phase: v[0],
-                    ocv: v[1],
-                    dcv: v[2],
-                    effects: v[3],
-                    active: false
+                system: {
+                    phase,
+                    ocv,
+                    dcv,
+                    effects,
+                    active: false,
+                    description: effects,
+                    XMLID,
                 }
             }
 
-            await HeroSystem6eItem.create(itemData, { parent: actor })
+            // if (attack) {
+            //     await makeAttack(item)
+            //     itemData.system.subType = "attack";
+            //     itemData.system.knockbackMultiplier = 1;
+            //     itemData.system.usesStrength = true;
+            //     itemData.system.LEVELS = { value: 0, max: 0 };
+            //     itemData.system.uses = 'ocv';
+            //     itemData.system.targets = 'dcv';
+            // }
+
+            let item = await HeroSystem6eItem.create(itemData, { parent: actor })
+            if (attack) {
+                await makeAttack(item)
+            }
         }
     }
 
@@ -391,23 +413,23 @@ export async function applyCharacterSheet(xmlDoc) {
     }
 
 
-    // Default Strike attack
-    let itemData = {
-        type: "attack",
-        name: "strike",
-        system: {
-            //xmlid: "HANDTOHANDATTACK",
-            knockbackMultiplier: 1,
-            usesStrength: true,
-            rules: "This is the basic attack maneuver",
-            XMLID: "HANDTOHANDATTACK",
-            NAME: "Strike",
-            LEVELS: { value: 0, max: 0 },
-            ALIAS: "HANDTOHANDATTACK",
-        }
+    // // Default Strike attack
+    // let itemData = {
+    //     type: "attack",
+    //     name: "strike",
+    //     system: {
+    //         //xmlid: "HANDTOHANDATTACK",
+    //         knockbackMultiplier: 1,
+    //         usesStrength: true,
+    //         rules: "This is the basic attack maneuver",
+    //         XMLID: "HANDTOHANDATTACK",
+    //         NAME: "Strike",
+    //         LEVELS: { value: 0, max: 0 },
+    //         ALIAS: "HANDTOHANDATTACK",
+    //     }
 
-    }
-    await HeroSystem6eItem.create(itemData, { parent: this.actor })
+    // }
+    // await HeroSystem6eItem.create(itemData, { parent: this.actor })
 
     await updateItemSubTypes(this.actor)
 
@@ -1798,7 +1820,7 @@ export async function makeAttack(item) {
 
     // Confirm this is an attack
     const configPowerInfo = getPowerInfo({ xmlid: xmlid, actor: item.actor })
-    if (!configPowerInfo || !configPowerInfo.powerType.includes("attack")) return
+    //if (!configPowerInfo || !configPowerInfo.powerType.includes("attack")) return
 
     let changes = {}
     //changes[`img`] = "icons/svg/sword.svg"
@@ -1806,7 +1828,7 @@ export async function makeAttack(item) {
 
     // Name
     let description = item.system.ALIAS
-    let name = item.system.NAME || description || configPowerInfo.xmlid
+    let name = item.system.NAME || description || configPowerInfo?.xmlid || item.system.name;
     changes[`name`] = name
 
     let levels = parseInt(item.system.LEVELS?.value) || parseInt(item.system.DC) || 0;
@@ -1872,6 +1894,10 @@ export async function makeAttack(item) {
     changes[`system.ocv`] = ocv
     changes[`system.dcv`] = dcv
     changes['system.stunBodyDamage'] = "stunbody"
+
+    // Maneuvers
+    if (item.system.type === "maneuver") {
+    }
 
     // ENTANGLE (not implemented)
     if (xmlid == "ENTANGLE") {
@@ -2038,10 +2064,6 @@ export async function makeAttack(item) {
         }
     }
 
-    // if (xmlid === "HANDTOHANDATTACK") {
-    //     await HeroSystem6eItem.create(itemData, { parent: this.actor })
-    //     return
-    // }
 
     if (xmlid === "HKA" || item.system.EFFECT?.indexOf("KILLING") > -1) {
         changes[`system.killing`] = true